requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
requests==2.31.0
python-dateutil==2.8.2
pillow==10.1.0
python-decouple==3.8
drf-yasg==1.21.7
django-filter==23.3
python-dotenv==1.0.0
matplotlib==3.7.1
numpy==1.26.2
pandas==2.1.4
